fix bug: Foreign Key Constraint not enforced when operating existing db file
Before, the "PRAGMA foreign_keys = ON" pragma was only executed when initializing a new db file. So when opening an existing db file and operate with it, no foreign key constraint will be enforced.
Now the pragma is moved into `fence` function, which will be executed every time db connection is established.