Change the algorithm of calculating return on shares
Add a new view: `share_trade_flows`, and make `share_trades` view use `share_trade_flows` view.
This changes the calculation method when a transaction involves two different non-standard assets, and one of them's change amount is 0 (this happens, for example, when a stock gives its dividend). In this case, the new algorithm uses the other account's change amount to calculate the value of flow, which is more accurate than simply deem the value of flow as 0.