Fix bug: `return_on_shares` view shows incorrect data if there is a stock split (or reverse stock split)
This bug was introduced by Commit 3807f2b
In commit 3807f2b, the algorithm to calculate the value of flow was changed to use the amount of change of the other account if one account in the transaction has an amount of change as 0 (this happens in, for example, stock dividend transactions). But this should only apply to transactions in which both accounts contain non-standard asset. If either account contains standard asset, the amount of change of standard asset should be used as the value of flow.
This commit also changes the field order of `check_absent_price` to be in accordance with prompt message.