Fix bug: wrongly detect an input line that is not the first line as a headline
The headline detection just examined if all the fields are not numbers, but did not check if this is the first line.
Now it is changed that only the first line can be detected as a headline.
This commit also changes the prompt better if a string is not a valid date where date string is required.